% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{repliapa_plot_allpas}
\alias{repliapa_plot_allpas}
\title{Plot all-PAS added value against a proximal/distal summary}
\usage{
repliapa_plot_allpas(comparison)
}
\arguments{
\item{comparison}{Output from \code{\link[=repliapa_compare_allpas_pd]{repliapa_compare_allpas_pd()}}.}
}
\value{
A ggplot object.
}
\description{
Plot all-PAS added value against a proximal/distal summary
}
