% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{repliapa_plot_calibration}
\alias{repliapa_plot_calibration}
\title{Plot guide-unit NTC calibration}
\usage{
repliapa_plot_calibration(calibration, max_alpha = 0.1)
}
\arguments{
\item{calibration}{A \code{repliapa_null_calibration} object.}

\item{max_alpha}{Largest nominal threshold displayed.}
}
\value{
A ggplot object.
}
\description{
Plot guide-unit NTC calibration
}
