% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{repliapa_plot_context}
\alias{repliapa_plot_context}
\title{Plot cross-context support and effect direction}
\usage{
repliapa_plot_context(comparison)
}
\arguments{
\item{comparison}{A \code{repliapa_context_comparison} object.}
}
\value{
A ggplot object.
}
\description{
Plot cross-context support and effect direction
}
