% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{repliapa_plot_reproducibility}
\alias{repliapa_plot_reproducibility}
\title{Plot frozen biological-replicate reproducibility}
\usage{
repliapa_plot_reproducibility(reproducibility, label_n = 6L)
}
\arguments{
\item{reproducibility}{Output from \code{\link[=repliapa_reproducibility]{repliapa_reproducibility()}}.}

\item{label_n}{Number of most significant development genes to label.}
}
\value{
A ggplot object.
}
\description{
Plot frozen biological-replicate reproducibility
}
